### 2024-10-09 #7339 @editorialbot
[PRE REVIEW]: PyLops-MPI - MPI Powered PyLops with mpi4py
**Submitting author:** @rohanbabbar04 (<a href="http://orcid.org/0000-0002-7203-7641">Rohan Babbar</a>)
**Repository:** https://github.com/PyLops/pylops-mpi
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.0
