+++ joss/paper.bib
@@ -106,14 +106,18 @@
 }
 
 @article{Nemeth:1999,
-  title={Least-squares migration of incomplete reflection data},
-  author={Nemeth, Tamas and Wu, Chengjun and Schuster, Gerard T},
-  journal={Geophysics},
-  volume={64},
-  number={1},
-  pages={208--221},
-  year={1999},
-  publisher={Society of Exploration Geophysicists}
+    author = {Nemeth, Tamas and Wu, Chengjun and Schuster, Gerard T.},
+    title = "{Least-squares migration incomplete reflection data}",
+    journal = {Geophysics},
+    volume = {64},
+    number = {1},
+    pages = {208-221},
+    year = {1999},
+    month = {02},
+    issn = {0016-8033},
+    doi = {10.1190/1.1444517},
+    url = {https://doi.org/10.1190/1.1444517},
+    eprint = {https://pubs.geoscienceworld.org/seg/geophysics/article-pdf/64/1/208/3173610/208.pdf},
 }
 
 @article{Ravasi:2022,
