### 2024-10-09 #7339 @danielskatz
@rohanbabbar04 - thanks for your submission. I'll be the track editor, working on it until it gets assigned to an editor, and then coming back at the end to handle final processing.

At this point, can you
1. ensure that references have spaces before them in the .md file, so that we don't have things like <img width="251" alt="Screenshot 2024-10-09 at 16 47 50" src="https://github.com/user-attachments/assets/aa82bb61-7849-488d-8be6-5a2896a26813"> where the space is missing in the PDF
2. Consider removing some of your paper in favor of pointing to documentation in the repo. Your paper is about 1800 words, and JOSS papers are recommended to be 250-1000 words.
3. Consider if you need an Acknowledgments section, to acknowledge any funding or other support or help

I'll mark this as paused - please ping me when you are ready to proceed or have questions.
