### 2024-10-10 3fcc294
Made changes to paper.md
