+++ joss/paper.md
@@ -27,55 +27,52 @@ bibliography: paper.bib
 
 # Summary
 
-Linear algebra operations and inverse problems represent the cornerstone of numerous algorithms in fields such as image
-processing, geophysics, signal processing, and remote sensing. This paper presents PyLops-MPI, an extension of the PyLops 
-framework, specifically designed to enable distributed computing in the solution of large-scale inverse problems in Python. 
-PyLops-MPI provides functionalities to parallelize any combination of PyLops operators with different reduction patterns, drop-in
-replacements for several PyLops operator that require changes in their inner working to be amenable to distributed computing, 
-and distributed solvers. By leveraging the Message Passing Interface (MPI) standard, the presented library can effectively unleash 
-the computational power of multiple nodes (or ranks), enabling users to efficiently scale their inversion problems with minimal 
-code modifications compared to their single-node equivalent PyLops codes.
+Linear algebra operations and inverse problems are fundamental to many algorithms in fields like image processing, geophysics, 
+signal processing, and remote sensing. This paper introduces PyLops-MPI, an extension of the PyLops framework designed for 
+distributed computing to solve large-scale inverse problems in Python. PyLops-MPI enables the parallelization of any combination 
+of PyLops operators with various reduction patterns, provides drop-in replacements for certain operators to facilitate distributed
+computing, and offers distributed solvers. By leveraging the Message Passing Interface (MPI) standard, this library effectively 
+harnesses the computational power of multiple nodes, allowing users to scale their inversion problems efficiently with minimal 
+changes to their single-node PyLops code.
 
 # Statement of need
 
-As scientific datasets grow and the demand for higher resolution increases, the need for distributed computing alongside
-matrix-free linear algebra becomes more critical. Nowadays, it is in fact common for the size of models and datasets to exceed 
-the memory capacity of a single machine—making it difficult to perform computations efficiently and accurately at the same time. 
-However, many linear operators used in scientific inverse problems can be usually decomposed in a series of computational blocks that, although being resource-intensive, can be effectively parallelized; this further emphasizes the necessity for a distributed computing approach to inverse problems.
+As scientific datasets grow and demand higher resolution, need for distributed computing and matrix-free linear algebra becomes essential. 
+Models and datasets now often exceed a single machine's memory, making efficient, accurate computation challenging. However, 
+many linear operators in scientific inverse problems can be decomposed into a series of computational blocks that, though 
+resource-intensive, are well-suited for parallelization, highlighting the need for a distributed approach.
 
 When addressing distributed inverse problems, we identify three distinct families of problems:
 
-- **1. Fully distributed models and data**: both model and data are distributed across nodes, with minimal
-  communication during the modeling process. Communication mainly occurs in the solver when dot
-  products or regularization terms (i.e., Laplacian) are applied. In this scenario, each node can easily
-  handle a portion of the model and data when applying the modelling operator and its adjoint.
+- **1. Fully distributed models and data**: Both model and data are distributed across nodes with minimal communication during 
+  modeling, mainly occurring in the solver for dot products or regularization (e.g., Laplacian). Each node handles a portion 
+  of the model and data when applying the operator and its adjoint.
 
-- **2. Distributed data, model available on all nodes**: in this case, data is distributed across nodes while the model is
-  available on all nodes. Communication is required during the adjoint pass when models produced by each node need
[…]
