+++ joss/paper.md
@@ -37,29 +37,25 @@ changes to their single-node PyLops code.
 
 # Statement of need
 
-As scientific datasets grow and demand higher resolution, need for distributed computing and matrix-free linear algebra becomes essential. 
-Models and datasets now often exceed a single machine's memory, making efficient, accurate computation challenging. However, 
-many linear operators in scientific inverse problems can be decomposed into a series of computational blocks that, though 
-resource-intensive, are well-suited for parallelization, highlighting the need for a distributed approach.
+As scientific datasets grow, the need for distributed computing and matrix-free linear algebra becomes crucial. 
+Models and datasets often exceed a single machine’s memory, making efficient computation challenging. Many linear operators 
+in scientific inverse problems can be decomposed into a series of computational blocks that are well-suited for parallelization, 
+emphasizing the need for a distributed approach.
 
 When addressing distributed inverse problems, we identify three distinct families of problems:
 
-- **1. Fully distributed models and data**: Both model and data are distributed across nodes with minimal communication during 
-  modeling, mainly occurring in the solver for dot products or regularization (e.g., Laplacian). Each node handles a portion 
-  of the model and data when applying the operator and its adjoint.
+- **1. Fully distributed models and data**: Both the model and data are split across nodes with minimal communication, mainly 
+  in the solver for dot products or regularization. Each node processes its own portion of the model and data.
 
-- **2. Distributed data, model available on all nodes**: In this case, data is distributed across nodes while the model is
-  available across all. Communication occurs during the adjoint pass when models produced by each node need
-  to be summed together, and in the solver for dot products on the data vector.
+- **2. Distributed data, model available on all nodes**: Data is distributed across nodes, but the model is available on all. 
+  Communication happens during the adjoint pass to sum models and in the solver for data vector operations.
 
-- **3. Model and data available on all nodes**: Here, communication is limited to the operator, with nodes having identical
-  copies of the data and model master. All nodes perform computations in the forward and adjoint passes of the operator, requiring
-  no communication in the solver.
+- **3. Model and data available on all nodes**: All nodes have identical copies of the data and model. Communication is limited 
+  to operator calculations, with no communication in solver needed.
 
 MPI for Python (mpi4py [@Dalcin:2021]) provides Python bindings for the MPI standard, allowing applications to leverage multiple 
-processors across workstations, clusters, and supercomputers. Recent updates (version 3.0 and above) have simplified usage and improved 
-data communication efficiency between nodes. Projects like mpi4py-fft [@Mortensen:2019], mcdc [@Morgan:2024], and mpi4jax [@mpi4jax] utilize mpi4py 
-to expand their distributed computing capabilities, improving efficiency and scalability.
+processors across workstations, clusters, and supercomputers. Projects like mpi4py-fft [@Mortensen:2019], mcdc [@Morgan:2024], and mpi4jax [@mpi4jax] 
+utilize mpi4py to expand their distributed computing capabilities, improving efficiency and scalability.
 
 PyLops-MPI, built on top of PyLops [@Ravasi:2020], leverages mpi4py to address large-scale problems in a distributed and parallel manner. It provides an 
 intuitive API for scattering and broadcasting data and models across nodes, allowing various mathematical operations (e.g., summation, subtraction, norms) 
@@ -80,37 +76,27 @@ The main components of the library include:
 
 ## DistributedArray
 
-The `pylops_mpi.DistributedArray` class serves as the fundamental array class used throughout the library, enabling 
[…]
