+++ joss/paper.md
@@ -27,43 +27,32 @@ bibliography: paper.bib
 
 # Summary
 
-Linear algebra operations and inverse problems are fundamental to many algorithms in fields like image processing, geophysics, 
-signal processing, and remote sensing. This paper introduces PyLops-MPI, an extension of the PyLops framework designed for 
-distributed computing to solve large-scale inverse problems in Python. PyLops-MPI enables the parallelization of any combination 
-of PyLops operators with various reduction patterns, provides drop-in replacements for certain operators to facilitate distributed
-computing, and offers distributed solvers. By leveraging the Message Passing Interface (MPI) standard, this library effectively 
-harnesses the computational power of multiple nodes, allowing users to scale their inversion problems efficiently with minimal 
-changes to their single-node PyLops code.
+Linear algebra and inverse problem theory are fundamental to many algorithms in signal processing, image processing, geophysics, and remote sensing. This paper introduces PyLops-MPI, an extension of the PyLops framework designed to enable distributed computing in the solution of large-scale inverse problems in Python. By leveraging the Message Passing Interface (MPI) standard, this library effectively 
+harnesses the computational power of multiple nodes, allowing users to scale their inverse problems efficiently with minimal 
+changes compared to their single-node PyLops code.
 
 # Statement of need
 
-As scientific datasets grow, the need for distributed computing and matrix-free linear algebra becomes crucial. 
-Models and datasets often exceed a single machine’s memory, making efficient computation challenging. Many linear operators 
-in scientific inverse problems can be decomposed into a series of computational blocks that are well-suited for parallelization, 
-emphasizing the need for a distributed approach.
+As scientific datasets grow and the demand for higher resolution increases, the use of distributed computing in matrix-free linear algebra becomes crucial. 
+Models and datasets can in fact easily exceed the memory of a single machine—making it difficult to perform computations efficiently and accurately at the same time. Nevertheless, many linear operators 
+in scientific inverse problems can be decomposed into a series of computational blocks that are well-suited for parallelization.
 
 When addressing distributed inverse problems, we identify three distinct families of problems:
 
-- **1. Fully distributed models and data**: Both the model and data are split across nodes with minimal communication, mainly 
-  in the solver for dot products or regularization. Each node processes its own portion of the model and data.
+- Both model and data are split across nodes, with each node processing its own portion of the model and data. This leads to minimal  
+  communication, mainly when performing dot products in the solver or in the regularization terms. 
 
-- **2. Distributed data, model available on all nodes**: Data is distributed across nodes, but the model is available on all. 
+- Data is distributed across nodes, whilst the model is available on all nodes. 
   Communication happens during the adjoint pass to sum models and in the solver for data vector operations.
 
-- **3. Model and data available on all nodes**: All nodes have identical copies of the data and model. Communication is limited 
-  to operator calculations, with no communication in solver needed.
+- All nodes have identical copies of the data and model. Communication only happens within 
+  the operator, with no communication in solver needed.
 
 MPI for Python (mpi4py [@Dalcin:2021]) provides Python bindings for the MPI standard, allowing applications to leverage multiple 
-processors across workstations, clusters, and supercomputers. Projects like mpi4py-fft [@Mortensen:2019], mcdc [@Morgan:2024], and mpi4jax [@mpi4jax] 
-utilize mpi4py to expand their distributed computing capabilities, improving efficiency and scalability.
-
[…]
