### 2024-10-11 a4cafab
Minor changes to paper.md
