### 2024-10-11 ee22336
Made changes to paper.md (#109)
* Made changes to paper.md

* Update paper.md

* More cleanup

* Small changes to bib file

* Minor changes to paper.md

---------

Co-authored-by: mrava87 <matteoravasi@gmail.com>
