+++ joss/paper.bib
@@ -1,5 +1,5 @@
 @article{Ravasi:2020,
-    author = {M. Ravasi and I. Vasconcelos},
+    author = {Ravasi, Matteo and Vasconcelos, Ivan},
     title = "{PyLops - A linear-operator Python library for scalable algebra and optimization}",
     journal = {SoftwareX},
     year = 2020,
@@ -28,7 +28,7 @@
     volume = {4},
     number = {36},
     pages = {1340},
-    author = {Mikael Mortensen and Lisandro Dalcin and David Elliot Keyes},
+    author = {Mortensen, Mikael and Dalcin, Lisandro and Keyes, David Elliot},
     title = {mpi4py-fft: Parallel Fast Fourier Transforms with MPI for Python},
     journal = {Journal of Open Source Software}
 }
@@ -128,7 +128,7 @@
    DOI={10.1109/tgrs.2022.3179626},
    journal={IEEE Transactions on Geoscience and Remote Sensing},
    publisher={Institute of Electrical and Electronics Engineers (IEEE)},
-   author={Ravasi, Matteo and Selvan, Tamin and Luiken, Nick},
+   author={Ravasi, Matteo and Selvan, Tamil and Luiken, Nick},
    year={2022},
    pages={1–14}
 }
