+++ joss/paper.md
@@ -15,7 +15,7 @@ authors:
     orcid: 0000-0002-0741-6602
     affiliation: 3
 affiliations:
-  - name: Computer Science and Engineering, Cluster Innovation Center, University of Delhi, Delhi, India.
+  - name: Cluster Innovation Center, University of Delhi, Delhi, India.
     index: 1
   - name: Earth Science and Engineering, Physical Sciences and Engineering (PSE), King Abdullah University of Science and Technology (KAUST), Thuwal, Kingdom of Saudi Arabia.
     index: 2
@@ -27,149 +27,83 @@ bibliography: paper.bib
 
 # Summary
 
-Linear algebra operations and inverse problems represent the cornerstone of numerous algorithms in fields such as image
-processing, geophysics, signal processing, and remote sensing. This paper presents PyLops-MPI, an extension of the PyLops 
-framework, specifically designed to enable distributed computing in the solution of large-scale inverse problems in Python. 
-PyLops-MPI provides functionalities to parallelize any combination of PyLops operators with different reduction patterns, drop-in
-replacements for several PyLops operator that require changes in their inner working to be amenable to distributed computing, 
-and distributed solvers. By leveraging the Message Passing Interface (MPI) standard, the presented library can effectively unleash 
-the computational power of multiple nodes (or ranks), enabling users to efficiently scale their inversion problems with minimal 
-code modifications compared to their single-node equivalent PyLops codes.
+Linear algebra and inverse problem theory are fundamental to many algorithms in signal processing, image processing, geophysics, and remote sensing. This paper introduces PyLops-MPI, an extension of the PyLops framework designed to enable distributed computing in the solution of large-scale inverse problems in Python. By leveraging the Message Passing Interface (MPI) standard, this library effectively 
+harnesses the computational power of multiple nodes, allowing users to scale their inverse problems efficiently with minimal 
+changes compared to their single-node PyLops code.
 
 # Statement of need
 
-As scientific datasets grow and the demand for higher resolution increases, the need for distributed computing alongside
-matrix-free linear algebra becomes more critical. Nowadays, it is in fact common for the size of models and datasets to exceed 
-the memory capacity of a single machine—making it difficult to perform computations efficiently and accurately at the same time. 
-However, many linear operators used in scientific inverse problems can be usually decomposed in a series of computational blocks that, although being resource-intensive, can be effectively parallelized; this further emphasizes the necessity for a distributed computing approach to inverse problems.
+As scientific datasets grow and the demand for higher resolution increases, the use of distributed computing in matrix-free linear algebra becomes crucial. 
+Models and datasets can in fact easily exceed the memory of a single machine—making it difficult to perform computations efficiently and accurately at the same time. Nevertheless, many linear operators 
+in scientific inverse problems can be decomposed into a series of computational blocks that are well-suited for parallelization.
 
 When addressing distributed inverse problems, we identify three distinct families of problems:
 
-- **1. Fully distributed models and data**: both model and data are distributed across nodes, with minimal
-  communication during the modeling process. Communication mainly occurs in the solver when dot
-  products or regularization terms (i.e., Laplacian) are applied. In this scenario, each node can easily
-  handle a portion of the model and data when applying the modelling operator and its adjoint.
+- **1. Fully distributed models and data**: Both model and data are split across nodes, with each node processing its own portion of the model and data. This leads to minimal  
+  communication, mainly when performing dot products in the solver or in the regularization terms. 
 
[…]
