### 2024-10-11 26a1ab4
Minor update in paper.md
