### 2024-10-11 905429c
Minor change in paper.md
