### 2024-10-11 #7339 @rohanbabbar04
> @rohanbabbar04 - thanks for your submission. I'll be the track editor, working on it until it gets assigned to an editor, and then coming back at the end to handle final processing.
>
> At this point, can you
>
>     1. ensure that references have spaces before them in the .md file, so that we don't have things like <img alt="Screenshot 2024-10-09 at 16 47 50" width="251" src="https://private-user-images.githubusercontent.com/2913845/375126758-aa82bb61-7849-488d-8be6-5a2896a26813.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3Mjg2MzI3MjksIm5iZiI6MTcyODYzMjQyOSwicGF0aCI6Ii8yOTEzODQ1LzM3NTEyNjc1OC1hYTgyYmI2MS03ODQ5LTQ4OGQtOGJlNi01YTI4OTZhMjY4MTMucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI0MTAxMSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNDEwMTFUMDc0MDI5WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9MzgxOTg2YTk1ZjdkZWExOWM0ZGQ5MDIyNzU5ODY0Njg0NTU1YmY2OTMxZmI4YTY2ZmQxNjJjNjU2MGY3NWFjNyZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.M9P3DR7Iq4_P3oXjdwPHKvZPK6wx5INo2IYwcGIRY7I"> where the space is missing in the PDF
>
>     2. Consider removing some of your paper in favor of pointing to documentation in the repo. Your paper is about 1800 words, and JOSS papers are recommended to be 250-1000 words.
>
>     3. Consider if you need an Acknowledgments section, to acknowledge any funding or other support or help
>
>
> I'll mark this as paused - please ping me when you are ready to proceed or have questions.

Thanks @danielskatz for a quick review of the paper and providing your feedback.
I have made the updates following your suggestions.
- Ensured that all references have spaces before them in the .md file to prevent any formatting issues in the PDF.
- The content is now within 250-1000 words to align the paper with JOSS's recommended length.
- Added an Acknowledgments section to acknowledge support.

Please let me know if there are any other changes required. I’m happy to proceed whenever you’re ready.
