### 2024-10-21 d67018b
feat: improved handling of shapes
