### 2024-10-21 3108060
minor: small code simplication
