### 2024-10-23 a2abc37
feat: added mask to DistributedArray
