### 2024-10-23 cc699df
feat: added mask to DistributedArray
