### 2024-10-23 5144312
doc: added example of using mask in DistributedArray
