### 2024-10-23 7a5ea19
test: added tests when using mask in DistributedArray
