+++ _sources/tutorials/sg_execution_times.rst.txt
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**02:20.088** total execution time for 2 files **from tutorials**:
+**02:23.297** total execution time for 3 files **from tutorials**:
 
 .. container::
 
@@ -33,8 +33,11 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_tutorials_poststack.py` (``poststack.py``)
-     - 02:16.968
+     - 02:18.327
      - 0.0
    * - :ref:`sphx_glr_tutorials_lsm.py` (``lsm.py``)
-     - 00:03.120
+     - 00:03.045
+     - 0.0
+   * - :ref:`sphx_glr_tutorials_mdd.py` (``mdd.py``)
+     - 00:01.925
      - 0.0
