+++ _sources/gallery/plot_stacked_array.rst.txt
@@ -403,7 +403,7 @@ of distributed arrays
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.015 seconds)
+   **Total running time of the script:** (0 minutes 0.014 seconds)
 
 
 .. _sphx_glr_download_gallery_plot_stacked_array.py:
