+++ _sources/gallery/plot_mdc.rst.txt
@@ -273,7 +273,7 @@ Finally let's display input model, data and adjoint model
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.480 seconds)
+   **Total running time of the script:** (0 minutes 0.477 seconds)
 
 
 .. _sphx_glr_download_gallery_plot_mdc.py:
