+++ _sources/tutorials/mdd.rst.txt
@@ -1,0 +1,406 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "tutorials/mdd.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_tutorials_mdd.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_tutorials_mdd.py:
+
+
+Multi-Dimensional Deconvolution
+===============================
+This example shows how to set-up and run a Multi-Dimensional Deconvolution
+problem in a distributed fashion, leveraging the :py:class:`pylops_mpi.waveeqprocessing.MDC`
+class.
+
+More precisely, compared to its counterpart in the PyLops documentation, this example distributes
+the frequency slices of the kernel of the MDC operator across multiple processes. Whilst both the
+entire model and data sit on all processes, within the MDC operator, and more precisely when the
+:py:class:`pylops_mpi.signalprocessing.Fredholm1` is called, different groups of frequencies are
+processed by the different ranks.
+
+.. GENERATED FROM PYTHON SOURCE LINES 15-34
+
+.. code-block:: Python
+
+
+    import numpy as np
+    from scipy.signal import filtfilt
+    from matplotlib import pyplot as plt
+    from mpi4py import MPI
+
+    from pylops.utils.seismicevents import hyperbolic2d, makeaxis
+    from pylops.utils.tapers import taper3d
+    from pylops.utils.wavelets import ricker
+
+    import pylops_mpi
+    from pylops_mpi.DistributedArray import local_split, Partition
+
+    plt.close("all")
+    rank = MPI.COMM_WORLD.Get_rank()
+    size = MPI.COMM_WORLD.Get_size()
+    dtype = np.float32
+    cdtype = np.complex64
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 35-37
+
+Let's start by creating a set of hyperbolic events to be used as
+our MDC kernel as well as the model
+
+.. GENERATED FROM PYTHON SOURCE LINES 37-104
+
+.. code-block:: Python
+
+
+    # Input parameters
+    par = {
+        "ox": -300,
+        "dx": 10,
+        "nx": 61,
+        "oy": -500,
+        "dy": 10,
+        "ny": 101,
+        "ot": 0,
+        "dt": 0.004,
+        "nt": 400,
+        "f0": 20,
+        "nfmax": 200,
+    }
+
+    t0_m = 0.2
+    vrms_m = 1100.0
+    amp_m = 1.0
+
+    t0_G = (0.2, 0.5, 0.7)
+    vrms_G = (1200.0, 1500.0, 2000.0)
+    amp_G = (1.0, 0.6, 0.5)
+
+    # Taper
+    tap = taper3d(par["nt"], (par["ny"], par["nx"]), (5, 5), tapertype="hanning")
+
+    # Create axis
+    t, t2, x, y = makeaxis(par)
+
+    # Create wavelet
+    wav = ricker(t[:41], f0=par["f0"])[0]
+
+    # Generate model
+    mrefl, mwav = hyperbolic2d(x, t, t0_m, vrms_m, amp_m, wav)
+
+    # Generate operator
+    G, Gwav = np.zeros((par["ny"], par["nx"], par["nt"])), np.zeros(
[…]

[…]
