### 2024-10-23 9e89e29
test: generalize tests to any choice of nranks
