### 2024-10-23 dd0dbaa
ci: Add nranks to matrix
