### 2024-10-23 a4406f9
ci: drop python3.8
