### 2024-10-23 69f03d3
test: fix some test to work with a variety of nranks
