### 2024-10-24 a018f21
test: increase rtol in test_linearop
