### 2024-10-24 0e5464d
test: increase rtol in remaining test_linearop
