+++ _sources/gallery/plot_mpilinop.rst.txt
@@ -104,7 +104,7 @@ with the same partitioning.
 
  .. code-block:: none
 
-    y: <DistributedArray with global shape=(242,), local shape=(242,), dtype=float64, processes=[0])> 
+    y: <DistributedArray with global shape=(np.int64(242),), local shape=(242,), dtype=float64, processes=[0])> 
 
 
 
@@ -214,7 +214,7 @@ to ensure that the operators align appropriately with their corresponding
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.344 seconds)
+   **Total running time of the script:** (0 minutes 0.341 seconds)
 
 
 .. _sphx_glr_download_gallery_plot_mpilinop.py:
