+++ _sources/gallery/plot_distributed_array.rst.txt
@@ -308,7 +308,7 @@ of the array and multiplies the corresponding elements together.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 1.483 seconds)
+   **Total running time of the script:** (0 minutes 1.481 seconds)
 
 
 .. _sphx_glr_download_gallery_plot_distributed_array.py:

[…]
