### 2024-11-14 1711229
feat: added proposal for UNSAFE_BROADCAST partition
