+++ _sources/tutorials/poststack.rst.txt
@@ -284,7 +284,7 @@ normal equations, and :math:`\mathbf{d}^{Norm}` is the data of the normal equati
         99        6.1544e-01         6.5026e-01     6.5026e-01
        100        6.1561e-01         6.4212e-01     6.4212e-01
 
-    Iterations = 100        Total time (s) = 29.64
+    Iterations = 100        Total time (s) = 29.21
     -----------------------------------------------------------------
 
 
@@ -349,7 +349,7 @@ normal equations, and :math:`\mathbf{d}^{Norm}` is the data of the normal equati
         99         6.3033e-01         1.9903e+02
        100         6.3026e-01         1.9777e+02
 
-    Iterations = 100        Total time (s) = 44.28
+    Iterations = 100        Total time (s) = 44.36
     -------------------------------------------------------
 
 
@@ -481,7 +481,7 @@ Finally, we display the modeling and inversion results
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (2 minutes 18.208 seconds)
+   **Total running time of the script:** (2 minutes 18.159 seconds)
 
 
 .. _sphx_glr_download_tutorials_poststack.py:
