+++ _sources/gallery/plot_cgls.rst.txt
@@ -170,7 +170,7 @@ function of the DistributedArray as shown below.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.175 seconds)
+   **Total running time of the script:** (0 minutes 0.173 seconds)
 
 
 .. _sphx_glr_download_gallery_plot_cgls.py:
