+++ _sources/gallery/plot_stacking.rst.txt
@@ -279,7 +279,7 @@ corresponding vector :math:`x`.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.429 seconds)
+   **Total running time of the script:** (0 minutes 0.456 seconds)
 
 
 .. _sphx_glr_download_gallery_plot_stacking.py:
