+++ _sources/gallery/plot_distributed_array.rst.txt
@@ -24,7 +24,7 @@ This example shows how to use the :py:class:`pylops_mpi.DistributedArray`.
 This class provides a way to distribute arrays across multiple processes in
 a parallel computing environment.
 
-.. GENERATED FROM PYTHON SOURCE LINES 8-22
+.. GENERATED FROM PYTHON SOURCE LINES 8-27
 
 .. code-block:: Python
 
@@ -39,6 +39,11 @@ a parallel computing environment.
     plt.close("all")
     np.random.seed(42)
 
+    # MPI parameters
+    size = MPI.COMM_WORLD.Get_size()  # number of nodes
+    rank = MPI.COMM_WORLD.Get_rank()  # rank of current node
+
+
     # Defining the global shape of the distributed array
     global_shape = (10, 5)
 
@@ -49,13 +54,13 @@ a parallel computing environment.
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 23-26
+.. GENERATED FROM PYTHON SOURCE LINES 28-31
 
-Let's start by defining the
-class with the input parameters ``global_shape``,
-``partition``, and ``axis``. Here's an example implementation of the class with ``axis=0``.
+Let's start by defining the class with the input parameters ``global_shape``,
+``partition``, and ``axis``. Here's an example implementation of the class
+with ``axis=0``.
 
-.. GENERATED FROM PYTHON SOURCE LINES 26-34
+.. GENERATED FROM PYTHON SOURCE LINES 31-39
 
 .. code-block:: Python
 
@@ -79,12 +84,12 @@ class with the input parameters ``global_shape``,
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 35-37
+.. GENERATED FROM PYTHON SOURCE LINES 40-42
 
 Below is an implementation to show how the global array is distributed along
 the second axis.
 
-.. GENERATED FROM PYTHON SOURCE LINES 37-45
+.. GENERATED FROM PYTHON SOURCE LINES 42-50
 
 .. code-block:: Python
 
@@ -108,7 +113,7 @@ the second axis.
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 46-51
+.. GENERATED FROM PYTHON SOURCE LINES 51-56
 
 You also have the option of directly including the ``local_shapes`` as a parameter
 to the :py:class:`pylops_mpi.DistributedArray`. This will enable the assignment
@@ -116,7 +121,7 @@ of shapes to local arrays on each rank. However, it's essential to ensure that
 the number of processes matches the length of ``local_shapes``, and that the
 combined local shapes should align with the ``global_shape`` along the desired ``axis``.
 
-.. GENERATED FROM PYTHON SOURCE LINES 51-59
+.. GENERATED FROM PYTHON SOURCE LINES 56-64
 
 .. code-block:: Python
 
@@ -140,14 +145,14 @@ combined local shapes should align with the ``global_shape`` along the desired `
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 60-64
+.. GENERATED FROM PYTHON SOURCE LINES 65-69
 
 To convert a random NumPy array into a ``pylops_mpi.DistributedArray``,
 you can use the ``to_dist`` classmethod. This method allows you to distribute
 the array across multiple processes for parallel computation.
 Below is an example implementation depicting the same.
 
-.. GENERATED FROM PYTHON SOURCE LINES 64-74
+.. GENERATED FROM PYTHON SOURCE LINES 69-79
 
 .. code-block:: Python
 
@@ -185,12 +190,15 @@ Below is an example implementation depicting the same.
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 75-77
[…]
