### 2024-11-23 c61d7af
feat: flush all prints in solver
