+++ _sources/gallery/plot_mpilinop.rst.txt
@@ -214,7 +214,7 @@ to ensure that the operators align appropriately with their corresponding
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.357 seconds)
+   **Total running time of the script:** (0 minutes 0.349 seconds)
 
 
 .. _sphx_glr_download_gallery_plot_mpilinop.py:
