### 2024-11-24 a36e19d
bug: fixed StackedDistributedArray.norm to work with cupy arrays
