### 2024-11-25 #7512 @mikaem
@gkthiruvathukal I have gone through the checklist and I have very few issues. The software installs immediately on my mac and the tests run with just a few fails that could simply be due to roundoff on my computer. The tests all pass on github actions and I see no need to follow this up further here. The documentation is good. One minor thing, you need to install the dev-environment.yml in order to run tests. That is not really what the README says and it should probably be fixed.

I have just a few issue with the written paper. The figure contains very small text and I don't think it is referenced in running text? I think it can safely be removed or else it needs to be introduced properly. Also, at the bottom of page 2 **DistributedArray** and **StackedDistributedArray** are formatted in bold, but I think they should be formatted like the other code objects, like `pylops_mpi.MPIVstack`.
