+++ _sources/tutorials/lsm.rst.txt
@@ -304,7 +304,7 @@ solver.
         99       -1.2546e-02         2.8528e+00     2.8528e+00
        100       -1.2389e-02         2.8342e+00     2.8342e+00
 
-    Iterations = 100        Total time (s) = 1.21
+    Iterations = 100        Total time (s) = 1.25
     -----------------------------------------------------------------
 
 
@@ -385,7 +385,7 @@ solver.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 3.165 seconds)
+   **Total running time of the script:** (0 minutes 3.227 seconds)
 
 
 .. _sphx_glr_download_tutorials_lsm.py:
