+++ _sources/gallery/plot_derivative.rst.txt
@@ -262,7 +262,7 @@ dimensions and concatenates them.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 1.329 seconds)
+   **Total running time of the script:** (0 minutes 1.308 seconds)
 
 
 .. _sphx_glr_download_gallery_plot_derivative.py:
