### 2024-11-26 b82223c
bug: allow unsafe_broadcast in VStack and Fredholm1
