+++ _sources/tutorials/mdd.rst.txt
@@ -367,7 +367,7 @@ back for our input model
         49       0.00e+00+0.00e+00j    3.7578e-01     3.7578e-01
         50       0.00e+00+0.00e+00j    3.6788e-01     3.6788e-01
 
-    Iterations = 50        Total time (s) = 1.36
+    Iterations = 50        Total time (s) = 1.11
     -----------------------------------------------------------------
 
 
@@ -377,7 +377,7 @@ back for our input model
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 2.227 seconds)
+   **Total running time of the script:** (0 minutes 1.951 seconds)
 
 
 .. _sphx_glr_download_tutorials_mdd.py:
