+++ _sources/gallery/plot_distributed_array.rst.txt
@@ -472,7 +472,7 @@ And finally with 2d-arrays distributed over axis=0
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 1.516 seconds)
+   **Total running time of the script:** (0 minutes 1.462 seconds)
 
 
 .. _sphx_glr_download_gallery_plot_distributed_array.py:
