### 2024-11-27 #7512 @rohanbabbar04
Hi @mikaem ,

Thank you for your feedback!

Here are some points regarding the feedback:

- The installation procedure for developers is documented on the [Installation](https://pylops.github.io/pylops-mpi/installation.html#id2) page. We typically don't include this information in the README, but instead, document it in the installation guide. Additionally, we provide a Makefile that we ask developers to use for installation.
- **DistributedArray** and **StackedDistributedArray** have been updated to `pylops_mpi.DistributedArray` and `pylops_mpi.StackedDistributedArray`.
- I would like to retain the figure, as it illustrates the main architecture of the library and operators. However, I agree that it should be referenced appropriately. I've added a line referencing Fig 1.
