### 2024-12-04 15a462e
bug: fix ncp.sum for cupy arrays not accepting lists
