### 2024-12-21 #7512 @rohanbabbar04
Hi @gkthiruvathukal ,

Here is the latest tag version - v0.2.1
DOI - [10.5281/zenodo.14539781](https://doi.org/10.5281/zenodo.14539781)
