### 2024-12-23 #7512 @rohanbabbar04
> Here is the latest tag version - v0.2.1 DOI - [10.5281/zenodo.14539781](https://doi.org/10.5281/zenodo.14539781)

@gkthiruvathukal, just let me know if you need anything else
