### 2025-01-07 #7512 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
