### 2025-01-07 be70251
minor changes in paper.bib
