+++ joss/paper.bib
@@ -29,7 +29,7 @@
     number = {36},
     pages = {1340},
     author = {Mortensen, Mikael and Dalcin, Lisandro and Keyes, David Elliot},
-    title = {mpi4py-fft: Parallel Fast Fourier Transforms with MPI for Python},
+    title = {mpi4py-fft: Parallel Fast {F}ourier Transforms with MPI for {P}ython},
     journal = {Journal of Open Source Software}
 }
 
@@ -42,7 +42,7 @@
     number = {96},
     pages = {6415},
     author = {Joanna Piper Morgan and Ilham Variansyah and Samuel L. Pasmann and Kayla B. Clements and Braxton Cuneo and Alexander Mote and Charles Goodman and Caleb Shaw and Jordan Northrop and Rohan Pankaj and Ethan Lame and Benjamin Whewell and Ryan G. McClarren and Todd S. Palmer and Lizhong Chen and Dmitriy Y. Anistratov and C. T. Kelley and Camille J. Palmer and Kyle E. Niemeyer},
-    title = {Monte Carlo / Dynamic Code (MC/DC): An accelerated Python package for fully transient neutron transport and rapid methods development},
+    title = {{M}onte {C}arlo / Dynamic Code (MC/DC): An accelerated {P}ython package for fully transient neutron transport and rapid methods development},
     journal = {Journal of Open Source Software}
 }
 
