### 2025-01-07 #7512 @danielskatz
👋 @rohanbabbar04 - as track chair, I've proofread this paper, and have suggested some small changes in https://github.com/PyLops/pylops-mpi/pull/125. Please merge this, or let me know what you disagree with, then we can continue to acceptance and publication.
